requirements.txt
mfrc522==0.0.7
RPi.GPIO==0.7.1
RPLCD==1.3.1
six==1.16.0
smbus2==0.4.3
spidev==3.6
